refact(Customers): reorganize redux files
