+++ src/sagas/index.ts
@@ -5,7 +5,7 @@ import balancesSaga      from './balances';
 import bankAccountsSaga  from 'services/bank-accounts/saga.js';
 import organizationsSaga from './organizations';
 import transactionsSaga  from './transactions';
-import customersSaga     from './customers';
+import customersSaga     from 'services/customers/saga.js';
 import categoriesSaga    from './categories';
 import membersSaga       from './members';
 import usersSaga         from './users';
