--- src/selectors/customers.js
@@ -1,4 +1,0 @@
-const selectCustomers       = (state) => state.customers.items;
-const selectCustomersStatus = (state) => state.customers.status;
-
-export { selectCustomers, selectCustomersStatus };
