+++ src/services/customers/index.ts
@@ -1,0 +1,3 @@
+export * from './types';
+export * from './actions.js';
+export * from './selectors.js';
