+++ src/services/customers/selectors.js
@@ -1,0 +1,13 @@
+const selectCustomers       = (state) => state.customers.items;
+const selectCustomersStatus = (state) => state.customers.status;
+
+const selectCustomer       = (state) => state.customer.data;
+const selectCustomerStatus = (state) => state.customer.status;
+
+export {
+  selectCustomers,
+  selectCustomersStatus,
+
+  selectCustomer,
+  selectCustomerStatus,
+};
