+++ src/services/customers/types.ts
@@ -1,0 +1,5 @@
+export interface Customer {
+  id:               number;
+  name:             string;
+  invoiceDetails:   string;
+}
